ml/requirements.txt
aniso8601==9.0.1
astroid==2.5.1
attrs==20.3.0
blinker==1.4
bson==0.5.10
certifi==2020.12.5
click==7.1.2
Deprecated==1.2.12
dnspython==2.1.0
Flask==1.1.2
flask-buzz==0.1.15
Flask-Cors==3.0.10
Flask-HTTPAuth==4.2.0
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-MySQLdb==0.2.0
flask-praetorian==1.1.0
Flask-PyMongo==2.3.0
flask-restplus==0.13.0
importlib-metadata==3.7.3
inflection==0.3.1
isort==5.7.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
jsonschema==3.2.0
lazy-object-proxy==1.5.2
MarkupSafe==1.1.1
marshmallow==3.10.0
mccabe==0.6.1
mysqlclient==2.0.3
parso==0.8.1
passlib==1.7.4
pendulum==2.1.2
psycopg2==2.8.6
py-buzz==1.0.3
PyJWT==1.7.1
pylint==2.7.2
pymongo==3.11.3
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pytzdata==2020.1
sentry-sdk==1.0.0
six==1.15.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.1
pandas
numpy
sklearn
os
sys
mysql.connector

